ObjectTracker
Overview:
The object tracker tracks an object based on color. We set thresholds for identifying colors by converting the whole image into a hue saturation value, then a binary mask is created which detects the object of only that color based on the threshold values we specified.

Dependencies:
OpenCV 2.4 or greater
imutils
Numpy which is also a dependency for Open CV

Running the code:
The program can be run through command prompt or through a suitable IDE. By default it captures the video through your webcam, but, you can also run it for a recorded video on your device, by running the program on command prompt by specifying suitable arguments(for ex: python Object-Tracker.py --video videos/example.mp4 , note: the directory should be the same as your Object-Tracker.py folder.